requirements.txt
flask==2.1.0
werkzeug==2.1.0
kafka-python==2.0.2
requests==2.26.0
